Exception handling - delete

+++ src/main/java/com/fundamentals/WebServiceJavaSpringBoot/services/UserService.java
@@ -2,8 +2,11 @@ package com.fundamentals.WebServiceJavaSpringBoot.services;
 
 import com.fundamentals.WebServiceJavaSpringBoot.entities.User;
 import com.fundamentals.WebServiceJavaSpringBoot.repositories.UserRepository;
+import com.fundamentals.WebServiceJavaSpringBoot.services.exceptions.DatabaseException;
 import com.fundamentals.WebServiceJavaSpringBoot.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,8 +31,15 @@ public class UserService  {
         return repository.save(obj);
     }
 
-    public void delete (Long id) {
-        repository.deleteById(id);
+    public void delete(Long id) {
+        try {
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
